Move dropdowns and burger into their own component files
